Move Slack Import to App Layer.
