+++ app/slackimport.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2015 Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-package api
+package app
 
 import (
 	"archive/zip"
@@ -16,7 +16,6 @@ import (
 	"unicode/utf8"
 
 	l4g "github.com/alecthomas/log4go"
-	"github.com/mattermost/platform/app"
 	"github.com/mattermost/platform/model"
 	"github.com/mattermost/platform/utils"
 )
@@ -138,7 +137,7 @@ func SlackAddUsers(teamId string, slackusers []SlackUser, log *bytes.Buffer) map
 
 	// Need the team
 	var team *model.Team
-	if result := <-app.Srv.Store.Team().Get(teamId); result.Err != nil {
+	if result := <-Srv.Store.Team().Get(teamId); result.Err != nil {
 		log.WriteString(utils.T("api.slackimport.slack_import.team_fail"))
 		return addedUsers
 	} else {
@@ -160,10 +159,10 @@ func SlackAddUsers(teamId string, slackusers []SlackUser, log *bytes.Buffer) map
 		password := model.NewId()
 
 		// Check for email conflict and use existing user if found
-		if result := <-app.Srv.Store.User().GetByEmail(email); result.Err == nil {
+		if result := <-Srv.Store.User().GetByEmail(email); result.Err == nil {
 			existingUser := result.Data.(*model.User)
 			addedUsers[sUser.Id] = existingUser
-			if err := app.JoinUserToTeam(team, addedUsers[sUser.Id]); err != nil {
+			if err := JoinUserToTeam(team, addedUsers[sUser.Id]); err != nil {
 				log.WriteString(utils.T("api.slackimport.slack_add_users.merge_existing_failed", map[string]interface{}{"Email": existingUser.Email, "Username": existingUser.Username}))
 			} else {
 				log.WriteString(utils.T("api.slackimport.slack_add_users.merge_existing", map[string]interface{}{"Email": existingUser.Email, "Username": existingUser.Username}))
@@ -192,7 +191,7 @@ func SlackAddUsers(teamId string, slackusers []SlackUser, log *bytes.Buffer) map
 
 func SlackAddBotUser(teamId string, log *bytes.Buffer) *model.User {
 	var team *model.Team
-	if result := <-app.Srv.Store.Team().Get(teamId); result.Err != nil {
+	if result := <-Srv.Store.Team().Get(teamId); result.Err != nil {
 		log.WriteString(utils.T("api.slackimport.slack_import.team_fail"))
 		return nil
 	} else {
@@ -245,7 +244,7 @@ func SlackAddPosts(teamId string, channel *model.Channel, posts []SlackPost, use
 			}
 			ImportPost(&newPost)
 			for _, fileId := range newPost.FileIds {
-				if result := <-app.Srv.Store.FileInfo().AttachToPost(fileId, newPost.Id); result.Err != nil {
+				if result := <-Srv.Store.FileInfo().AttachToPost(fileId, newPost.Id); result.Err != nil {
 					l4g.Error(utils.T("api.slackimport.slack_add_posts.attach_files.error"), newPost.Id, newPost.FileIds, result.Err)
 				}
 			}
@@ -413,7 +412,7 @@ func SlackUploadFile(sPost SlackPost, uploads map[string]*zip.File, teamId strin
 }
 
 func deactivateSlackBotUser(user *model.User) {
-	_, err := app.UpdateActive(user, false)
+	_, err := UpdateActive(user, false)
 	if err != nil {
 		l4g.Warn(utils.T("api.slackimport.slack_deactivate_bot_user.failed_to_deactivate", err))
 	}
@@ -424,7 +423,7 @@ func addSlackUsersToChannel(members []string, users map[string]*model.User, chan
 		if user, ok := users[member]; !ok {
 			log.WriteString(utils.T("api.slackimport.slack_add_channels.failed_to_add_user", map[string]interface{}{"Username": "?"}))
 		} else {
-			if _, err := app.AddUserToChannel(user, channel); err != nil {
+			if _, err := AddUserToChannel(user, channel); err != nil {
 				log.WriteString(utils.T("api.slackimport.slack_add_channels.failed_to_add_user", map[string]interface{}{"Username": user.Username}))
 			}
 		}
@@ -474,7 +473,7 @@ func SlackAddChannels(teamId string, slackchannels []SlackChannel, posts map[str
 		mChannel := ImportChannel(&newChannel)
 		if mChannel == nil {
 			// Maybe it already exists?
-			if result := <-app.Srv.Store.Channel().GetByName(teamId, sChannel.Name); result.Err != nil {
+			if result := <-Srv.Store.Channel().GetByName(teamId, sChannel.Name); result.Err != nil {
 				l4g.Warn(utils.T("api.slackimport.slack_add_channels.import_failed.warn"), newChannel.DisplayName)
 				log.WriteString(utils.T("api.slackimport.slack_add_channels.import_failed", map[string]interface{}{"DisplayName": newChannel.DisplayName}))
 				continue
@@ -609,7 +608,7 @@ func SlackImport(fileData multipart.File, fileSize int64, teamID string) (*model
 		deactivateSlackBotUser(botUser)
 	}
 
-	app.InvalidateAllCaches()
+	InvalidateAllCaches()
 
 	log.WriteString(utils.T("api.slackimport.slack_import.notes"))
 	log.WriteString("=======\r\n\r\n")
